+++ starboard/raspi/shared/main.cc
@@ -16,6 +16,7 @@
 #include <time.h>
 
 #include "starboard/configuration.h"
+#include "starboard/event.h"
 #include "starboard/raspi/shared/application_dispmanx.h"
 #include "starboard/shared/signal/crash_signals.h"
 #include "starboard/shared/signal/debug_signals.h"
@@ -54,11 +55,27 @@ int main(int argc, char** argv) {
     third_party::crashpad::wrapper::InstallCrashpadHandler(
         true, ca_certificates_path);
   }
-#endif
+#endif  // SB_IS(EVERGREEN_COMPATIBLE)
+
+#if SB_MODULAR_BUILD
+  return SbRunStarboardMain(argc, argv, SbEventHandle);
+#else
   starboard::raspi::shared::ApplicationDispmanx application;
   int result = application.Run(argc, argv);
   starboard::shared::signal::UninstallSuspendSignalHandlers();
   starboard::shared::signal::UninstallDebugSignalHandlers();
   starboard::shared::signal::UninstallCrashSignalHandlers();
   return result;
+#endif  // SB_MODULAR_BUILD
 }
+
+#if SB_MODULAR_BUILD
+int SbRunStarboardMain(int argc, char** argv, SbEventHandleCallback callback) {
+  starboard::raspi::shared::ApplicationDispmanx application(callback);
+  int result = application.Run(argc, argv);
+  starboard::shared::signal::UninstallSuspendSignalHandlers();
+  starboard::shared::signal::UninstallDebugSignalHandlers();
+  starboard::shared::signal::UninstallCrashSignalHandlers();
+  return result;
+}
+#endif  // SB_MODULAR_BUILD