+++ starboard/shared/starboard/application.cc
@@ -44,7 +44,11 @@ void Dispatch(SbEventType type, void* data, SbEventDataDestructor destructor) {
   SbEvent event;
   event.type = type;
   event.data = data;
+#if SB_MODULAR_BUILD
+  Application::Get()->sb_event_handle_callback_(&event);
+#else
   SbEventHandle(&event);
+#endif  // SB_MODULAR_BUILD
   if (destructor) {
     destructor(event.data);
   }
@@ -65,11 +69,22 @@ volatile SbAtomic32 g_next_event_id = 0;
 
 Application* Application::g_instance = NULL;
 
+#if SB_MODULAR_BUILD
+Application::Application(SbEventHandleCallback sb_event_handle_callback)
+    : error_level_(0),
+      thread_(SbThreadGetCurrent()),
+      start_link_(NULL),
+      state_(kStateUnstarted),
+      sb_event_handle_callback_(sb_event_handle_callback) {
+  SB_CHECK(sb_event_handle_callback_)
+      << "sb_event_handle_callback_ has not been set.";
+#else
 Application::Application()
     : error_level_(0),
       thread_(SbThreadGetCurrent()),
       start_link_(NULL),
       state_(kStateUnstarted) {
+#endif  // SB_MODULAR_BUILD
   Application* old_instance =
       reinterpret_cast<Application*>(SbAtomicAcquire_CompareAndSwapPtr(
           reinterpret_cast<SbAtomicPtr*>(&g_instance),
@@ -543,7 +558,11 @@ bool Application::HandleEventAndUpdateState(Application::Event* event) {
 // OnSuspend() is called after SbEventHandle(kSbEventTypeSuspend).
 #endif  // SB_API_VERSION >= 13
 
+#if SB_MODULAR_BUILD
+  sb_event_handle_callback_(scoped_event->event);
+#else
   SbEventHandle(scoped_event->event);
+#endif  // SB_MODULAR_BUILD
 
 #if SB_API_VERSION >= 13
   switch (scoped_event->event->type) {