+++ starboard/shared/starboard/queue_application.h
@@ -35,7 +35,12 @@ namespace starboard {
 // manage event dispatching.
 class QueueApplication : public Application {
  public:
+#if SB_MODULAR_BUILD
+  explicit QueueApplication(SbEventHandleCallback sb_event_handle_callback)
+      : Application(sb_event_handle_callback) {}
+#else
   QueueApplication() {}
+#endif  // SB_MODULAR_BUILD
   ~QueueApplication() override {}
 
  protected: