+++ starboard/shared/x11/application_x11.cc
@@ -696,12 +696,19 @@ int ErrorHandler(Display* display, XErrorEvent* event) {
 
 using shared::starboard::player::filter::CpuVideoFrame;
 
+#if SB_MODULAR_BUILD
+ApplicationX11::ApplicationX11(SbEventHandleCallback sb_event_handle_callback)
+#else
 ApplicationX11::ApplicationX11()
+#endif  // SB_MODULAR_BUILD
     : wake_up_atom_(None),
       wm_delete_atom_(None),
 #if SB_API_VERSION >= 13
       wm_change_state_atom_(None),
 #endif  // SB_API_VERSION >= 13
+#if SB_MODULAR_BUILD
+      QueueApplication(sb_event_handle_callback),
+#endif  // SB_MODULAR_BUILD
       composite_event_id_(kSbEventIdInvalid),
       display_(NULL),
       paste_buffer_key_release_pending_(false) {