+++ starboard/CHANGELOG.md
@@ -31,6 +31,14 @@ specific to the stream (i.e. don't change per access unit) to
 ### Renamed SbPlayerWriteSample2() to SbPlayerWriteSamples()
 To better reflect the fact that it may write multiple samples in one call.
 
+### Added SB_MODULAR_BUILD for supporting modular builds
+This configuration is set for modular builds, which have:
+  1. Application binary built as a shared library.
+  2. Either of a) or b)
+        a. Starboard built at a shared library and a separate loader_app executable.
+        b. A loader_app executable with Starboard built in ( Evergreen ).
+
+
 ## Version 14
 ### Add MP3, FLAC, and PCM values to SbMediaAudioCodec.
 This makes it possible to support these codecs in the future.